refactored CUDA, Metal and oneAPI backends into extensions. Refactored tests into simpler files. Added oneAPI backend. Added tests for extension finding.
