+++ docs/Project.toml
@@ -1,5 +1,6 @@
 [deps]
+Atomix = "a9b6321e-bd34-4604-b9c9-b65b8de01458"
 Documenter = "e30172f5-a6a5-5a46-863b-614d45cd2de4"
 
 [compat]
-Documenter = "1.7"
+Documenter = "1"
