+++ docs/make_local.jl
@@ -1,0 +1,5 @@
+# Activate docs environment and use ("develop") local library
+using Pkg
+Pkg.activate(@__DIR__)
+Pkg.develop(path=joinpath(@__DIR__, ".."))
+include("make.jl")
