--- lib/AtomixCUDA/.gitignore
@@ -1,2 +1,0 @@
-Manifest.toml
-
