--- lib/AtomixCUDA/Project.toml
@@ -1,13 +1,0 @@
-name = "AtomixCUDA"
-uuid = "6171a885-8764-404f-b64d-f8edb1679b6f"
-authors = ["Takafumi Arakaki <aka.tkf@gmail.com> and contributors"]
-version = "1.0"
-
-[deps]
-Atomix = "a9b6321e-bd34-4604-b9c9-b65b8de01458"
-CUDA = "052768ef-5323-5732-b1bb-66c8b64840ba"
-
-[compat]
-julia = "1.6"
-Atomix = "1"
-CUDA = "5"
