--- lib/AtomixCUDA/test/AtomixCUDATests/src/AtomixCUDATests.jl
@@ -1,7 +1,0 @@
-module AtomixCUDATests
-
-include("utils.jl")
-include("test_core.jl")
-include("test_sugar.jl")
-
-end  # module AtomixCUDATests
