--- lib/AtomixCUDA/test/AtomixCUDATests/src/utils.jl
@@ -1,13 +1,0 @@
-module Utils
-
-using CUDA
-
-function cuda(f)
-    function g()
-        f()
-        nothing
-    end
-    CUDA.@cuda g()
-end
-
-end  # module
