--- lib/AtomixCUDA/test/Project.toml
@@ -1,6 +1,0 @@
-[deps]
-Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
-TestFunctionRunner = "792026f5-ac9a-4a19-adcb-47b0ce2deb5d"
-
-[compat]
-TestFunctionRunner = "0.1"
