--- lib/AtomixCUDA/test/runtests.jl
@@ -1,2 +1,0 @@
-using TestFunctionRunner
-TestFunctionRunner.@run
