--- lib/AtomixMetal/Project.toml
@@ -1,13 +1,0 @@
-name = "AtomixMetal"
-uuid = "fc0d88fc-5db5-40d7-bf35-da84fd759def"
-authors = ["Andrei-Leonard Nicusan <a.l.nicusan@gmail.com> and contributors"]
-version = "1.0"
-
-[deps]
-Atomix = "a9b6321e-bd34-4604-b9c9-b65b8de01458"
-Metal = "dde4c033-4e86-420c-a63e-0dd931031962"
-
-[compat]
-julia = "1.6"
-Atomix = "1"
-Metal = "1"
