--- lib/AtomixMetal/test/AtomixMetalTests/Project.toml
@@ -1,13 +1,0 @@
-name = "AtomixMetalTests"
-uuid = "03c00b03-d8de-4838-97ce-ef69879af46e"
-authors = ["Andrei-Leonard Nicusan <a.l.nicusan@gmail.com> and contributors"]
-version = "0.1.0-DEV"
-
-[deps]
-Atomix = "a9b6321e-bd34-4604-b9c9-b65b8de01458"
-AtomixMetal = "fc0d88fc-5db5-40d7-bf35-da84fd759def"
-Metal = "dde4c033-4e86-420c-a63e-0dd931031962"
-Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
-
-[compat]
-julia = "1.6"
