--- lib/AtomixMetal/test/AtomixMetalTests/src/AtomixMetalTests.jl
@@ -1,7 +1,0 @@
-module AtomixMetalTests
-
-include("utils.jl")
-include("test_core.jl")
-include("test_sugar.jl")
-
-end  # module AtomixMetalTests
