--- lib/AtomixMetal/test/AtomixMetalTests/src/utils.jl
@@ -1,13 +1,0 @@
-module Utils
-
-using Metal
-
-function metal(f)
-    function g()
-        f()
-        nothing
-    end
-    Metal.@metal g()
-end
-
-end  # module
