--- test/AtomixTests/Project.toml
@@ -1,13 +1,0 @@
-name = "AtomixTests"
-uuid = "6f95a7dd-85ce-4dbe-a121-27757241c193"
-authors = ["Takafumi Arakaki <aka.tkf@gmail.com> and contributors"]
-version = "0.1.0-DEV"
-
-[deps]
-Atomix = "a9b6321e-bd34-4604-b9c9-b65b8de01458"
-Documenter = "e30172f5-a6a5-5a46-863b-614d45cd2de4"
-Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
-UnsafeAtomics = "013be700-e6cd-48c3-b4a1-df204f14c38f"
-
-[compat]
-julia = "1"
