--- test/AtomixTests/src/test_doctest.jl
@@ -1,11 +1,0 @@
-module TestDoctest
-
-import Atomix
-using Documenter: doctest
-using Test
-
-function test_doctest()
-    doctest(Atomix; manual = false)
-end
-
-end  # module
