+++ test/Project.toml
@@ -1,8 +1,3 @@
 [deps]
-Documenter = "e30172f5-a6a5-5a46-863b-614d45cd2de4"
+Pkg = "44cfe95a-1eb2-52ea-b672-e2afdf69b78f"
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
-TestFunctionRunner = "792026f5-ac9a-4a19-adcb-47b0ce2deb5d"
-UnsafeAtomics = "013be700-e6cd-48c3-b4a1-df204f14c38f"
-
-[compat]
-UnsafeAtomics = "0.2"
